requirements.txt
numpy
pandas
streamlit
streamlit_webrtc

scikit-learn
sentence_transformers

av

python-dotenv
SpeechRecognition 
google.generativeai
moviepy 